+++ OpenRTM_aist_v121e_paho_mqtt_interface/OpenRTM_aist_paho_mqtt_module/InPortPahoSubscriber.py
@@ -4,7 +4,7 @@
 ##
 # @file   InPortPahoSubscriber.py
 # @brief  InPortPahoSubscriber class
-# @date   2020/12/09
+# @date   2020/12/10
 # @author Daishi Yoshino
 #
 # Copyright (C) 2020
@@ -20,7 +20,6 @@ import OpenRTM__POA,OpenRTM
 import time
 import sys
 from OpenRTM_aist_paho_mqtt_module.paho_client.PahoSubscriber import PahoSubscriber
-from OpenRTM_aist.ManagerActionListener import ManagerActionListeners
 
 ##
 # @class InPortPahoSubscriber
@@ -56,9 +55,6 @@ class InPortPahoSubscriber(OpenRTM_aist.InPortProvider, PahoSubscriber):
   # @brief Destructor
   #
   def __del__(self):
-    print("[disconnecting from MQTT broker start]")
-    PahoSubscriber.paho_disconnect(self)
-    print("[disconnecting from MQTT broker end]")
     PahoSubscriber.__del__(self)
     return
 
@@ -358,7 +354,7 @@ class InPortPahoSubscriber(OpenRTM_aist.InPortProvider, PahoSubscriber):
 # @class ManagerActionListener
 # @brief ManagerActionListener class
 #
-class ManagerActionListener(ManagerActionListeners):
+class ManagerActionListener:
   def __init__(self, InPortPahoSubscriber):
     self._InPortPahoSubscriber = InPortPahoSubscriber
 
@@ -369,7 +365,9 @@ class ManagerActionListener(ManagerActionListeners):
   # @brief Clean up mqtt communication module instance when RTC exit
   #
   def postShutdown(self):
-    self._InPortPahoSubscriber.__del__()
+    print("[disconnecting from MQTT broker start]")
+    self._InPortPahoSubscriber.paho_disconnect()
+    print("[disconnecting from MQTT broker end]")
 
   def preReinit(self):
     pass