+++ OpenRTM_aist_v121e_paho_mqtt_interface/OpenRTM_aist_paho_mqtt_module/OutPortPahoPubSecure.py
@@ -4,7 +4,7 @@
 ##
 # @file  OutPortPahoPubSecure.py
 # @brief OutPortPahoPubSecure class
-# @date   2020/12/09
+# @date   2020/12/10
 # @author Daishi Yoshino
 #
 # Copyright (C) 2020
@@ -21,7 +21,6 @@ import RTC
 import time
 import sys
 from OpenRTM_aist_paho_mqtt_module.paho_client.PahoPubSecure import PahoPubSecure
-from OpenRTM_aist.ManagerActionListener import ManagerActionListeners
 
 ##
 # @class OutPortPahoPubSecure
@@ -49,9 +48,6 @@ class OutPortPahoPubSecure(OpenRTM_aist.InPortConsumer, PahoPubSecure):
   #
   def __del__(self, CorbaConsumer=PahoPubSecure):
     self._rtcout.RTC_PARANOID("~OutPortPahoPubSecure()")
-    print("[disconnecting from MQTT broker start]")
-    PahoPubSecure.paho_disconnect(self)
-    print("[disconnecting from MQTT broker end]")
     PahoPubSecure.__del__(self)
     return
 
@@ -460,7 +456,7 @@ class OutPortPahoPubSecure(OpenRTM_aist.InPortConsumer, PahoPubSecure):
 # @class ManagerActionListener
 # @brief ManagerActionListener class
 #
-class ManagerActionListener(ManagerActionListeners):
+class ManagerActionListener:
   def __init__(self, OutPortPahoPubSecure):
     self._OutPortPahoPubSecure = OutPortPahoPubSecure
 
@@ -471,7 +467,9 @@ class ManagerActionListener(ManagerActionListeners):
   # @brief Clean up mqtt communication module instance when RTC exit
   #
   def postShutdown(self):
-    self._OutPortPahoPubSecure.__del__()
+    print("[disconnecting from MQTT broker start]")
+    self._OutPortPahoPubSecure.paho_disconnect()
+    print("[disconnecting from MQTT broker end]")
 
   def preReinit(self):
     pass